backend/requirements.txt
flask
flask-sqlalchemy
flask_cors
waitress
python-dotenv